Cursorによるリファクタリング
