+++ src/constants/storage/index.ts
@@ -1,0 +1,4 @@
+export const STORAGE_KEYS = {
+  DECK_CARDS: "deckCards_k",
+  DECK_NAME: "deckName_k",
+} as const;
