+++ src/types/deck/index.ts
@@ -1,0 +1,6 @@
+import type { Card } from "../card";
+
+export interface DeckCard {
+  readonly card: Card;
+  count: number;
+}
