+++ src/types/filter/index.ts
@@ -1,0 +1,8 @@
+import type { CardKind, CardType } from "../card";
+
+export interface FilterCriteria {
+  text: string;
+  kind: CardKind[];
+  type: CardType[];
+  tags: string[];
+}
